Adding functionality to HealthVariable
HealthVariables now need base values for internal parametrization, as well as
a list of which internal attributes can be updated by external models. Helper
functions to output a list of all updatable parameters, to know the lenght of
said list, and to update internal parameters, are also now intrinsic to this
class.