Creating new class to store HealthVariables
We now have a new class that manages a collection of HealthVariables. Everything
in this class MUST be a HealthVariable, other than the 'updatable' attribute.
This class behaves exactly like a HealthVariable, since a set of HVs is also an
HV. This means that this class inherits the same methods as HV, with a few
modifications to account for everything being an HV and make the process of
storing HVs more stream-lined.