Adding '+=' functionality between HVCollections
Now, two HealthVariable collections can be added with '+' and '+=' as well.
However, keep in mind that the standard fields for the type corresponding to the
first one in the operation will change, so, if you create a new instance of that
same type, it will come with these fields set to a default value of None.