Additional clean-up in post_init
In the HealthVariableCollection, if a field is set to None, it is not removed
after initialization, since it should not be part of the Collection. This fixes
the issue that '+' and '+=' operators change the standard fields of their type
model, and new instances of the same type/model take in the default fields as
defined in their class.