New helper to provide template for hidden vector
The ECM class now has a helper function to provide a specified number of copies
of the template of hidden vector it expects in subsequent function calls. This
can be used, for example, to provide different copies of equivalent hidden
states that will be evolved differently/separately by different model/estimators.