Need to add a CellModel to base
Decided to change the name to CellModel, seeing as it does not model health, but
only uses it to evolve the system and make voltage predictions. All the types in
the abstract methods are wrong, and will be fixed in a future commit. We need
this class in order to properly define the ECM model.