Added notebook to demonstrate ECM use
In this notebook, we look at a simple Rint model, a simple Thenevin model, and a
complete PNGV model, with series capacitor, two RC elements, and linear current
behavior. We demonstrate how to use these for updating the physical transient
hidden state, and how to calculate terminal voltage. We show how these models can
generate synthetic data. However, we did not change the A-SOH variables for this,
as it is just a demonstration. Additionally, we will streamline the data
generation inside the ECM model soon.