Minimally functioning ECM ready (#15)
* Streamlined base class for health variables

* Remove cov from setup.cfg

Fixes PyCharms debugger on windows

* Rename to HealthVariable, implement iterator

* Add tests for the new HealthVariable

* Implement function to update specific variables

Both individually or several at once

* Remove Test from class name to avoid confusing pytest

* Add unit tests for "iterate over everything"

* Reorganizing ECM and adding utilities

Made ECM a directory to better organize things in there. Also included SOC
interpolation class in utilities.

* Added ECM components

Main ECM components have been included.

* Changing base_values to take np.ndarray

In the new definition of HealthVariable, anything numeric must be passed as a
single number or as a np.ndarray. Therefore, I had to prevent base_values from
accepting a List, and enforce a float or np.ndarray instead.

* Testing interpolation capabilities

Added tests to make sure we can 1) infer SOC from base values and 2) perform
basic linear, quadratic, cubic interpolations.

* Adding constant value test case

* Adding A-SOH definition to ECM

Added class to represent A-SOH of an ECM, with defaults not requiring series
capacitor nor any RC elements. Also defaults to unit-valued Coulombic efficiency
and zero-valued hysteresis. Required fields are capacity, OCV, and series
resistance.

* Preparing tests for ECM A-SOH

Removed mention of HealthModel from ecm.__init__.py as it was causing issues in
running test. Added a conftest.py to help with things the test(s) should need.
Started putting together an A-SOH test.

* More things to full scale ECM A-SOH test

Added a few different levels of ECM A-SOH for us to test, but have not yet
implemented any of the tests.

* Tried to run a test but need more functionality

Tried to run a simple test, but realized functionality is not yet implemented in
the base class. Also, we need to change the way 'updatable' field is stored, so
that there exists an intrinsic order in the HealthVariable.

* Including HealthModel back in the base

We need to define a HealthModel for use in the ECM.

* Need to add a CellModel to base

Decided to change the name to CellModel, seeing as it does not model health, but
only uses it to evolve the system and make voltage predictions. All the types in
the abstract methods are wrong, and will be fixed in a future commit. We need
this class in order to properly define the ECM model.

* Added GeneralContainer

Also defined HiddenVector and AdvancedStateOfHealth, seeing as these are required
but the CellModel.

* Creating inputs for ECM

Temperature defaults to None, which should not trigger GeneralContainer's
to_numpy() method.

* Defining ECM output container

Added a TODO, but will reiterate here: recall that, if we want to denoise some
values (SOC, Qt, R0), we will likely need to include more fields, maybe in
new child classes defined specifically for the estimators (UKF, SPF, etc).

* minor changes to aesthetic

* Defined ECM transient vector class

Requires SOC and hysteresis, and accomodates for one or none series capacitor
charge, as well as none or multiple currents through resistors in RC elements.

* Added terminal voltage calculation to ECM

ECM now has the capability of calculating terminal voltage.

* Function to help with ECMTransient initialization

Implemented a function that can provide copies of an ECMTransientVector template
based on the ECM description. Can also help with initialization if specified
attributes/fields are provided.

* Moving ECMASOH to its own file

This way, the init can focus on the model, which will be quite large.

* Adding somewhat realistic OCV to utils

This way, when we need a template, we can easy get one.

* Changes to capacity returns

Now, 'value' property returns capacity in Amp-seconds, and the 'amp_hour'
property will return it in Amp-hour.

* Specifying capacity returns in Amp-hour

* Function to provide ECMASOH template

Created a function to help with providing a template/default for ECMASOH. It
cannot be used for unitialization, as some values as hard-coded.

* Specifying how to initialize ECM

ECM cell model now has the proper __init__ function.

* Setting q0 default to None

Now that the GeneralContainer checks for Number and Sized, we can set its default
to None, which is better than an empty list, seeing as this quantity either
exists or not, and should not be a list.

* Changes to GeneralContainer for consistency

Seeing as we had agreed that all numerical values should be stored as plain
numbers.Number types (such as float) or numpy.ndarrays (and not tuples or lists),
I modified the general container to account for that.

* Reflecting changes in General container

The general container now assumers all numerical values are numbers.Number or
numpy.ndarray, so I've modified the definition of the i_rc currents to reflect
that.

* Including helper function to solve hysteresis

Added a helper function to help solve for hysteresis under the assumption that
the current does not change sign during the time interval. This makes it easier
to deal with situations where that happens, as we can treat them as two separate
intervals.

* ECM can now update transient state

ECM now is able to provide a new transient state estimate.

* Need to allow arbitrary types in ECMTransient

* Allowing arbitrary types in GeneralContainer

Recall we expect children of this class to take in numpy.ndarrays, so we need to
allow for that.

* Realistic fake OCV now returns a numpy array

* Changing C0 to have a static default

* Fixing bug in GeneralContainer

the to_numpy method had a property decorator, which is not the intended behavior.
Additionally, inside it, the relevant_vals variable was initialized to a type,
instead of an empty tuple, which was causing issues when adding things to it.

* Terminal voltage calculator can use stored ECMInput

* Fixing bug in hysteresis update

In cases where current_k == 0.0, the division by its absolute value could cause
isues

* Fixed a bug in providing transient template

We now defined i_rc as a numpy array, so I had to change the way we provided
a template/initialization for the transient state.

* Minor modifications to transient update

Streamlined it as to not have to do the same calculation twice.

* Fixed bug in transient updated

Realized I was using resistance value as the RC time constant... Fixed!

* Changed default ASOH

Now template OCV uses cubic interpolation, which provides a smoother profile.

* Added notebook to demonstrate ECM use

In this notebook, we look at a simple Rint model, a simple Thenevin model, and a
complete PNGV model, with series capacitor, two RC elements, and linear current
behavior. We demonstrate how to use these for updating the physical transient
hidden state, and how to calculate terminal voltage. We show how these models can
generate synthetic data. However, we did not change the A-SOH variables for this,
as it is just a demonstration. Additionally, we will streamline the data
generation inside the ECM model soon.

* Minor changes to ECM demo

* Flake8 fixes

* Polishing many things up

---------

Co-authored-by: lward <lward@anl.gov>