ECM Dynamics and Simulator fully separate (#20)
* Fixes to the provide template functionality

The method had the old name, and used the old field names (capitalized). That has
been rectified.

* Additional bug fixes

RC components were still using lower case field names.

* Making main functions static methods + bug fixes

Since the transient update and the terminal voltage calculation no longer use
internal parameters (as there are none!), it makes sense for these to be static
methods (we no longer need to instantiate an ECM object that doesn't store
anything). Minor bug fixes were also performed (recall Coulombic efficiency was
its own HealthVariable).

* We now have an ECM simulator

This new class will take care of storing transient states, inputs, outputs, etc.
It can be used as a synthetic data generator as well as an ECM simulator.

* ECM simulator can now  keep its history

Added the possibility of keeping the full history in the simulator, including
inputs, transient states, and measurements. We are not keeping the A-SOH history
and for now, that does not change.

* Showing off simulator

Notebook now demonstrates how to use the ECM simulator.

* Adding markdown explanations

Moved comment explanations to markdown cells for nicer rendering.

* Removing init function

Since this class is not meant to be instantiated, and nothing happens upon
initialization, there is no need for an __init__ function.

* No more useless attributes

Previously, I was storing attributes that were not used anywhere else within the
class. That is no longer the case. This commit also comes with minor sytling
corrections.

* Removed inline comments

Before, I was using inline comments to explain what was happening in the update
of hysteresis. These have been moved to their own comment blocks.

* small linting change

* Fixing interpolator when base_vals not float

Before, if we were able to set an integer to base values post instantiation,
but that would lead to problems when interpolating. Taht behavior has been fixed.

* Fixing R0 issue

Previously, if R0 base values where a Number, it would ignore temperature
dependence when get_value was called. Now, instead of immediately returning when
base values is a float, we set the reference value to that and to the normal
temperature update.

* Fixing bug in hysteresis integrator

When the current behavior is set to 'constant', we should not check if the
current changes sign during the delta_t interval. Adding that condition to the
if-statement.

* Fixing hysteresis update

I had forgotten to adjust the amount of time of the second interval in the case
of the current changing sign during the delta_t interval.

* Fixing error on history after update

When updating, I change the transient vector, among other things. Since the
history pointed to that object, that was problematic, as past objects were being
changed. That has been fixed by making sure we always store model copies, rather
than the actual object, in history.

* Adding unit tests for ECM physics and simulation

ECM physics and simulator and now fully tested. We check for current integration
(both 'constant' and 'linear'), voltage drop across R0, voltage drop across C0,
hysteresis update, as well as asymptotic exponential decay in RC element.