Defining basic structure for online estimators
Here, we define the base classes for online estimators. We include a base
MultivariateRandomVariable, which is the common underlying description of hidden
states and outputs (maybe inputs too, but not clear on that yet). We also include
a class to help interface the CellModels, which expect specific types of inputs,
with the estimators, which should only look at these new classes define here
(these new classes should be all numeric, based on np.ndarrays)