Base definitions for online estimators (#23)
* Name change

Changed HiddenVector to TransientVector to avoid confusion when dealing with
hidden states during filtering estimators (where the name hidden will be used
explicitly to indicate other things, which include transient states as well as
other metrics).

* Changing name of outputs

Changing name from OutputMeasurements to OutputQuantities, as to avoid confusion
when we are dealing with measurements in the context of a filtering technique.

* Defining basic structure for online estimators

Here, we define the base classes for online estimators. We include a base
MultivariateRandomVariable, which is the common underlying description of hidden
states and outputs (maybe inputs too, but not clear on that yet). We also include
a class to help interface the CellModels, which expect specific types of inputs,
with the estimators, which should only look at these new classes define here
(these new classes should be all numeric, based on np.ndarrays)