Initial implementation of UKF (#26)
* New class for control variables

Just trying to make it generic, but can be changed later to only be the input
object to the cell model.

* Changing property to method in RandomVariable

Since many MultivariateRandomVariables will be initialized by providing a mean,
it doesn't make sense for that to be a separate property that could conflict with
a pydantic field.

* Name change for clarity

* Defining multivariate Gaussian

Including validators to make sure mean and covariance has matching dimensions.

* Moving definitions

Making more general definitions in __init__, and more specific ones in base.

* Cleaning up Kalman init

* Adding num_parameters and num_outputs to classes

HiddenState and OutputMeasurements now must have an easy way to get their
dimensionality

* Dimensions stored in the model-filter interface

* Changed expected value implementation in random var

MultivariateRandomVariable comes with a property expected_value that searches for
the mean and, if it cannot find it defined, raises a ValueError

* Model-filter interface should take in np.ndarrays

To be more explicit, we make it so that the model-estimator interface must take
in numpy.ndarrays, which is all the filters/estimators should have access to and
work with.

* Model-estimator interface clarification

Making it clearer the model-estimator interface should take a collection of
hidden states, and maybe other relevant parameters, as inputs to their main
functions.

* Reorganizing structure

Since there may be several "base" estimators (Kalman, UKF, SPF, etc.), it makes
sense to keep a "base" directory where they are all independently defined.

* Adding utilities to online estimator

* Utilities should be inside base.

* Adds control arg to the model-estimator interface

* Hidden update needs past and present controls

The update to the hidden state needs both past and present controls. The hidden
states provided correspond to those of the previous controls, and we need to
estimate what they would be at the time of the current control.

* Model needs control to predict outputs

* Wrote step 1 of UKF

The full step 1 (estimation) of the UKF has been implemented. We need to write
step 2 (correction), as well as all the necessary tests.

* UKF now has correction step

Implemented the correction step to the UKF.

* Fixing name from 'base' to 'general'

* Fixing bug in the model validator

Multivariate Gaussian had a typo in the model validator when checking that the
dimensions of the covariance matched those from the mean. That has been fixed.

* Typo in error message fixed

* First UKF demonstration working

Added a demonstration of how to use the UKF with a constant A-SOH Thenevin ECM.

* Deleting old notebook and rerunning

Re-ran UKF demonstration with much worse initialization of transient state to
demonstrate it still converges.

* Fixing bug with innovation

Innovation array needed to be flattened in order for matrix multiplication with
gain matrix to work properly.

* Adding initial state and control to OnlineEstimator

All instances of OnlineEstimator will likely need an initial state and control
variables, so it's wise to include them in the base class.

* Adding battery-independent UKF demo

Added a Lorenz chaotic system to the UKF demo, which will be used for the unit
test of the UKF.

* Adding Lorenz test to UKF

* Making Lorenz test more lenient

Given the very chaotic nature of the Lorenz system, we need to make the test a
bit more lenient, since the UKF cannot always keep up.

* More leniency on Lorenz test

* Typo on Lorenz test

* Adding seeded RNG to avoid tests failing on server

Tests seemed to always fail on the server side, despite always passing locally.
Added seeded RNG to circumvent this nuisance.

* Forgot to include the RNG in a couple of lines

* Simplifying MultivariateRandom

MultivariateRandomDistribution had annoying checks to return the first moment of
the distribution. Now, we just have a single abstractmethod to get the mean.

* Re-introducing get_mean to multivariate gaussian

* Demonstration is closer to the test

---------

Co-authored-by: victorventuri <victorventuri95@gmail.com>