Interface initialized in joint estimator separately
We decided to create a helper class to initialize the model-estimator interface,
as the basic init functionality of the joint estimator is the same, the main
different being which interface is used. In order to better streamline this
process, it makes sense to have new joint estimator initialize their interface
separately, but be initialized in the same manner.