Freezing ECMASOH template
Before, the ECMASOH provide_template returned an ASOH objects with a few inner
attributes marked as updatable, even though the top-level updatable set was empty.
This would cause problems when marking one component of the RC-pair as updatable,
since it would also automatically mark the other one as well. Now, to prevent this
issue, we return an ECMASOH object with all fields and inner fields frozen/fixed.