Interpolated HV had issues when array size of 1
When we update the health variables, the new base values are always numpy arrays,
even when the original base value was a float. These arrays are always shaped,
meaning that, in the cases where the original base value was a float, it is now
updated to an array of shape (1,). This is problematic for the interpolation
function. Therefore, we add a field validator to the base_values field to make
sure that, when an array is passed and has length of 1, we extract that value,
instead of storing the full array.