Issues with from_numpy update
The GeneralContainer from_numpy update is unable to distinguish between a field
that is supposed to be just a float and a field that is supposed to be an array,
but happens to have length of 1. By adding the validate_assignment flat to the
class definition, pydantic will be able to convert the values appropriately.