Adding capability to normalize joint state
Due to discrepancies in the order of magnitude of the joint state, numerical
conditioning of the estimation can be poor. In an attempt to mitigate this issue,
this commit implements functionality to be able to "normalize" the joint state
based on the initial values provided, so that the estimator hopefully only
operates on arrays composed of numbers close to 1 or 0.