Making sure A-SOH values never go below zero
This commit is a bad hotfix to an annoying problem: in many instances, the UKF
estimators will try to sample A-SOH parameters that are negative. This is a
problem for most ECM based HealthVariables, so I introduced here a quick check
to convert any negative value into a very small positive value. However, for
other models, it is possible that A-SOH parameters should be negative, which
would cause a major headache. So, this needs to me moved elsewhere!