Incomplete demonstration of Joint UKF
Here, we were finally able to run the joint UKF for over 10 minutes without
hiccups. However, this required a bad hotfix to the A-SOH parameters.
Additionally, the joint estimation is extremely slow... in 10-12 minutes, it only
goes over 10 hours of cycling, which is incredibly inneficient. This needs to be
investigate and fixed as soon as possible.