Added simple non-degrading Rint demo
Added a simple example in which we are trying to joint estimate a Rint model in
which the parameters do not degrade over time. Efficiency is still catastrophic,
and parts of the estimation seem to be problematic, especially the hysteresis
terms. But at least we have a full visualization of about 50 hours of simulation,
which is better than what we had with the Thenevin. This also had to be keyboard
interrupted, so there's that too...