Batching GeneralContainer
This commit adds the functionality necessary to batch GeneralContainer objects.
Now, all variables passed to the GeneralContainer (which were always either
floats or numpy arrays) are internally converted to a new object, a
BatchedVariable, which helps getting batch size and inner dimensions of each
variable. The appropriate conversion is dependent on if the variable is annotated
as a single-valued (SingleVal) variable, or as a multi-valued one (MultiVal).
These new Annotations were also added, and should help making transient states,
inputs, and outputs batched.