Allow batched calculations of physics models (#40)
* First ideas on batching

* Another pass on batching GeneralContainer

In this notebook, I take a simpler approach to try to allow batching within the
GeneralContainer class.

* Trying out GeneralContainer batching with annotations

* More flair to annotated validation

Just making sure the shapes of the arrays are adequate.

* Ignoring .DS_Store

I don't know why this only showed up now, but it's time to go...

* Batching GeneralContainer

This commit adds the functionality necessary to batch GeneralContainer objects.
Now, all variables passed to the GeneralContainer (which were always either
floats or numpy arrays) are internally converted to a new object, a
BatchedVariable, which helps getting batch size and inner dimensions of each
variable. The appropriate conversion is dependent on if the variable is annotated
as a single-valued (SingleVal) variable, or as a multi-valued one (MultiVal).
These new Annotations were also added, and should help making transient states,
inputs, and outputs batched.

* Moving BatchedVariable to utils

Before, we had issues with circular imports. By defining it in utils, we no
longer need to import things from base in utils, and from utils in base.

* Making ECMTransientVector batchable

These modifications make sure the ECM parameters are all batchable

* Including None as a possible BatchedVariable

If a value of None is passed, then a BatchedVariable object is returned with None
as the batched_values, and batch_size and inner_dimensions both set to 0.

* Converting optional transient inputs to BatchedVariable

Since everything in the GeneralContainer is assumed to be a BatchedVariable
object, we needed to make sure the optional inputs, originally just set to None,
are now BatchedVariables of None.

* Fixing optional fields for ECMInput

Similar to what we did with transient, Optional fields now have to be appropriate
BatchedVariable(batched_value=None)

* Removing Optional description

The Optional flag allows users to pass None to the variables, and skips
validation for some reason... By removing it and providing an appropriate
default, we can make sure that, if a value of None is passed, it will be
correctly converted to a BatchedVariable

* Removing Optional from imports

* Adding None to annotated single and multi-valued variables

* SingleVal stores a number if array is passed

In cases where a numpy array is passed with shapes (,), (1,), or (1,1), we
modify the code to store a single value (array of shape (,)) to avoid problems
with numpy broadcasting, since something with no dimensions (a plain number) is
always broadcastable.

* ECM handles new GeneralContainer-type objects

Modifications were made to reflect the fact that all attributes in
GeneralContainer are now BatchedVariable objects.

* Rerunning notebook to make sure it still works

* Adding copy method to BatchedVariable

Because the main batched_values in the object may be None, a Number, or an array,
I greated a method here to simply get the property (so that outside classes and
functions don't need to know if the value is None, for example)

* Adapting test to new BatchedVariables

Now, we need to compare the values to the .copy() method in BatchedVariable.

* Adding None as possible inpute type in siumlator

* Tests to make sure batching from init and updating work

These tests make sure we can batch from individual 1D and 2D arrays, and also
makes sure the from_numpy method can still properly work as intended.

* A few more notebook test and reruns, nothing major

* Replace BatchedVariable with a vanilla np.ndarray

* Use similar "2d numpy" concept in GeneralContainer

* Make the components of the ECM model batchable with numpy

* Use a list parameter for capacitance interp points

* Fix the initializer for ECMTransient

Was not respecting the 2d numpy array dictate

* Fix the up the tests

* Run the joint ukf test with CI

---------

Co-authored-by: victorventuri <victorventuri95@gmail.com>