Online revamp (#58)
* Starting re-structure

* Delta distribution has covariance equal 0

* Defining the minium requirements for a base filter

* Specifying utility functions for Kalman family

* Moving dimensionality specification to base class

* Changing import location

* Changing argument names for initialization for consistency

* Definition of UKF logic

We defined the UKF logic and general UKF filter to be used by the estimators.

* Changed import location

* Bug fixes

Forgot to specify default value for array1 in compute_unscented_covariance and
used wrong argument name in estimation_update for _predict_outputs

* Testing base UKF

Using a simple Lorenz model that operates directly on numpy arrays.

* Testing covariance assembly

* removing print statement

* Passing control values as numpy arrays

This way, the model wrapper knows to expect all arguments to be provided as
numpy arrays

* Defining ModelWrapper base class

* Removing CellModel from ModelWrapper

The base ModelWrapper should have no concept of the model being used.

* Chaning order of step outputs

Step needs to output estimated state, and then predicted output, not the other
way around.

* Performing a simple test of a single UKF step

* Introduce conversion between CellModel and ModelWrapper (#56)

* Add tests for the partial transients (#52)

* Add tests for the partial transients

* Add a documentation page about estimators

* Make the simulator more general, link to toolkit (#49)

* Make the simulator more general

* Add function to export to dataframe

* Flake8 fixes, test improvement

* Output from simulator to batdata format

Fixes #27

* Update notebooks to use more tests

* Add doc page for simulator

* Minor typo in docs

* Fix passthrough for updatable_transients kwargs

* Implement serialization to and from json (#53)

* Implement serialization to and from json

* Add version module

* Make ECMASOH serializable

* It seems the validator doesn't run if the field is optional

So, I just made it required because we give a default value
which has the intended effect

* Example CellModel->numpy function interface

* Finish documenting the input classes

* Add output to model interface

* Refactor in anticipation of a the DualUKF

* Use the new ModelFilter interface

* name change for clarity

* Raising NotImplementedError in base class

* allowing me to mess around with a notebook

* Easy way to convert between model-quantities and filter-quantities

Introduces a new function that takes a model-related object (which cannot be
a HealthVariable!), as well as its uncertainty (as a numpy array) and converts it
to a corresponding filter-related MultivariateRandomDistribution: if uncertainty
was provided, assumes a multivariate Gaussian; otherwise, assumes a delta
distribution.

* adding documentation for new function

* Defininig joint estimation framework

First pass at a general definition of the joint estimation framework.

* Refactor the online estimator class

* Forcing methods to be abstractmethods

The step and get_estimated_state methods need to be present in all children, so
we've made them abstractmethods. Additionally, the cell model is now stored as
'cell_model' explicitly.

* Defining how to assemble a state and get model-related objects

* Adding additional options (noise, hyperparameters)

The helper function to initialize the joint UKF now accepts the process and
sensor noise matrices, as well as UKF tuning parameters (alpha, beta, kappa)

* Fixing bug in the creation of cell model inputs

Previously, the joint wrapper would not actually update the transient vector, but
only copy what it had stored into a new array. This faulty behavior is fixed now.

* Imports were wrong

Imports have now been fixed.

* Rerun notebook with new joint implementation

Notebook has been rerun with the new joint implementation. However, it seems
we are still having some numerical issues with matrices not being positive
semi-definite.

* Update toolkit interface to new OnlineEstimator (#57)

Also introduce a `parameterize` to the test so
we can test new estimator types more easily

* Conversion model <=> filter objects no longer raises warnings

Before, the coversion was raising a warning because the mean being passed was
not flattened, but that is no longer the case.

* New JointUKF tested!

* Adapting the cell model wrapper tests to new code

* Removing unnecessary attribute

* Streamlining model <=> filter object conversion

---------

Co-authored-by: Logan Ward <WardLT@users.noreply.github.com>
Co-authored-by: Logan Ward <ward.logan.t@gmail.com>