Fix ECM broadcasting issues (#66)
* Fixing broadcasting issues

In numpy, the '+=' operation does not follow broadcasting rules, as it performs
the operation in place, rather than creting a copy. Therefore, we had to
substitute all of them.

* Fixing broadcasting issues

For interpolated variables, a number of issues popped up when the batch sizes
were not the same. Some of them are fixed, hopefully all, but not sure. Also, as
in the previous commit, '+=' is not broadcastable, so I had to fix that here too.

* Adding an appropriate batch test

Here, I add a test that mix-and-matches batched and un-batched transient vectors
and A-SOH, to make sure all operations run as intended.

* Making sure second dimension is correct in batched tests

* Optimizing numpy operations

There were a number of copies of numpy arrays that were not necessary. We have
changed it so that we only copy things which are changing to store them in a
different variable, but if it is part of an operation, the copy has been removed.