Adding clarity to non-interpolation case
In the case where no interpolation is needed, but the SOCInterpolatedHealth is
batched (batch size > 1), the input should only contain a number, so we check if
that was provided as a plain number, or as an array with dimensions. We return
the values following the criteria provided: if the input SOC was a number, we
return a 1d array, otherwise, we return a 2D array.