Fix a few more ECM broadcast issues (#67)
* Fixing broadcasting issues

In numpy, the '+=' operation does not follow broadcasting rules, as it performs
the operation in place, rather than creting a copy. Therefore, we had to
substitute all of them.

* Fixing broadcasting issues

For interpolated variables, a number of issues popped up when the batch sizes
were not the same. Some of them are fixed, hopefully all, but not sure. Also, as
in the previous commit, '+=' is not broadcastable, so I had to fix that here too.

* Adding an appropriate batch test

Here, I add a test that mix-and-matches batched and un-batched transient vectors
and A-SOH, to make sure all operations run as intended.

* Making sure second dimension is correct in batched tests

* Optimizing numpy operations

There were a number of copies of numpy arrays that were not necessary. We have
changed it so that we only copy things which are changing to store them in a
different variable, but if it is part of an operation, the copy has been removed.

* Fixing vestigial broadcasting issues

A few '+=' and '*=' operations still remained, and have been addressed. An
additional issues with the interpolator returning a 3D array when the soc inputs
were batched with shape (1,1) has also been fixed. Issues might still exist for
truly batched SOC inputs...

* Adding clarity to non-interpolation case

In the case where no interpolation is needed, but the SOCInterpolatedHealth is
batched (batch size > 1), the input should only contain a number, so we check if
that was provided as a plain number, or as an array with dimensions. We return
the values following the criteria provided: if the input SOC was a number, we
return a 1d array, otherwise, we return a 2D array.