Dual estimation framework done (#71)
* Defined DegradationModelInterface

The wrapper needed to interact with the A-SOH only filter in the dual estimation
framework is prepared. It does not take a degradation model yet, but that is to
come.

* Just creating the place where dual framework lives

* Changing name for consistency

* Finalized orignal code for cell and degradation specific wrappers

Completed the minimum necessary wrappers for dual estimation framework.

* Making it clearer means have to be flattened

Multivariate random variables require flat mean arrays. Here, we just make sure
that is the case.

* Fixing bug due to change in argument name

* Adding a way for multivariate random variables to be combined

Included the requirement that multivariate random variables can be easily
combined if they are of the same type (that is, of the same distribution, but
not necessarily equal PDFs, or even equal dimensions!)

* CellModel does not need to know which parts of A-SOH are updatable

* Finished first pass of full Dual framework implementation

Dual estimation framework fully implemented, but not yet tested. It also provides
a dual UKF to facilitate use.

* Changing order in which arguments are passed

Initialization of UKF should take critical parameters first, important parameters
second, and tuning/hyper- parameters last. That has been fixed.

* Making some attributes cached properties

This way, it is easier for initialization with incomplete set of parameters can
can be re-set post initialization, but prior to estimation steps.

* Facilitating type-hinting for dual UKF initialization

* Fixing bug in A-SOH initialization

A-SOH initialization to multivariate gaussian in the provide UKF helper function
had a bug: it passed the wrong argument to the function. That has been fixed.

* Fixes A-SOH values not being passed as np.ndarray

* Fixes bug related to returning None

Operations being performed return None, but altered the object. Therefore, they
could not be performed on the return statement of the helper function.

* First steps in demonstrate dual UKF, currently broken

* Fixing bug with combining MvRD

* First succesful run of Dual UKF

Numerical problems still exist in the Thenevin example, and need to be further
investigated. The performance on the Rint model is still poor.

* Rerunning notebook

Making sure the merge from main (which includes necessary fixes to broadcasting)
did not break anything.

* Allowing temperature to be None

If a value is given to temperature, it triggers the OCVEntropic part, which we
don't want to have. Allowing values of temperature to not be provided and set to
None avoids that problem.

* Adding a test for dual UKF

* simple rerun, nothing different

* Removing temperature from control names

Since we made temperature an optional transient attribute, and it is not provided
in our tests, we should remove it from the tuple of expected control names.

* Moving logic of tuning parameters

Tuning parameters used by filters have been moved to adequate locations.

* Allowing base model classes to create copies

HealthVariable and GeneralContainer can now create copies of themselves with
values from specified numpy array.

* Adding simple test for distributions

* Adding test for dual wrappers