First pass at adding convertors to dual-wrappers
Dual wrappers now have the capability of dealing with conversion operators from
the base ModelWrapper. They have been tested with the default conversion
(identity), and still work, though they are slower. Next, we will do the same to
the joint, make sure it works with identity convertors, and then test the proper
normalization.