Estimator normalization (#73)
* Creating transformations to MvRD

Specify base and Linear transformations to be applied to MvRD.

* MvRDs are now able to transform themselves

Require MvRDs to be able to take a specific transformation operator and output a
new copy of themselves, but with properly transformed values. This must not
change the existing values in the instance of the MvRD!

* Renaming conversion operators

In order to avoid the word 'transformation', which has implicit strict
mathematical meaning, I replaced it by 'conversion', which is softer. Mostly,
this avoids the problem of a linear transformation being different from a linear
function (does not obey additivity [f(a+b)=f(a)+f(b)] nor homogeneity
[f(c*a)=c*f(a)] conditions)

* Making nomenclature changes consistent

Making sure there is little to no reference to 'transformation' anymore, instead
using 'conversion' as much as possible.

* changing name to conversions

* Fixing matrix multiplication

Before, I was performing matrix multiplication with np.dot, which is incorrect.
It needs to use np.matmul.

* changing name from points to samples

* Allowing multiplicative array to be 0d

Added checks to make sure operations still work.

* Chaning nomenclature from points to samples

This makes it clearer what it is for.

* Fixing bug in inverse covariance

The check on the dimensionality of the multiplicative array was wrong.

* Fixing bug with validator

Needed to make sure to return multiplicative array after the validator tests

* fixing bug with conversion in delta

* Adding tests for the conversion

Making sure the conversion works by itself, as well as when coupled with the
multivariate random variables we have.

* typo fix

* Adding basic identity operator

Now, we have a simple identity operator that should not bring much overhead when
used.

* testing identity conversion operator

* Removed unnecessary buggy code

There was some leftover code from when we used to set the filter attributes
after their instantiation. Not only that is no longer needed, it was also wrong.

* Conversion operators passed to wrapper

Now, conversion operators can be passed to the ModelWrappers. The default is
the identity conversion (that is, no conversion).

* Adding auxiliary functions to base wrapper

* Making it clearer what is model and what is filter lingo

* typo in function name

* Fixing issues with inverse transform in wrapper

Wrapper was using wrong keyword for its inverse transforms. That has been fixed.

* First pass at adding convertors to dual-wrappers

Dual wrappers now have the capability of dealing with conversion operators from
the base ModelWrapper. They have been tested with the default conversion
(identity), and still work, though they are slower. Next, we will do the same to
the joint, make sure it works with identity convertors, and then test the proper
normalization.

* Joint wrapper now also accepts conversion operators

Added the conversion operator functionality to the Joint wrapper. Retested with
default (identity) operators, still works, thankfully! Now, on to normalization.

* Adds equal test but with normalization enabled

* Allowing None to go through identity operator

* Joint estimator has normalization of A-SOH

* rerun; still problematic

* Reran joint demo with smaller alpha

By using a smaller alpha parameters, sigma points are sampled closer to their
mean, and, therefore, we avoid issues with unphysical A-SOH parameters.

* Reran dual with smaller alpha

To see if the issues with the thenevin example also stemmed from unphysicallities
in the A-SOH, reran this notebook with smaller alpha for A-SOH UKF. It was able
to run for longer, but still ended up problematic... Let's see if that reamins
after normalization is added to the dual framework.

* Small changes for merge

Fixed tests that had broken as well.

* Remove unnecessary "Optional" types

---------

Co-authored-by: Logan Ward <WardLT@users.noreply.github.com>